Change site configuration for lastmod parameter
If gitInfo is enabled in the site configuration, Hugo uses the git Author Date for the lastmod value by default. I've changed it so that Hugo now takes the lastmod frontmatter param as the default, if it's set. If it isn't set, Hugo uses the git Author Date as a fallback.